Update Encasulation
private - this means that the information is only accessible within the class itself.

What happens in the house is only known by the people living there.
default - this is also called package-private, as we have seen so far. This means it is only accessible within the package.

You and the apartment managers know how much rent you pay a month, but no one else.
protected - this means that the information is only accessible by everything in the same package or subclasses.

You know who comes to your apartment, but so might the neighbors or complex managers... creepy.
public - this means that the classes and methods are accessible everywhere in our application. Therefore accessible by anything.

Everyone can go to the grocery store.